Move CoreAuditLogger and getAuditLogEvent into lib/auditLog folder
