+++ packages/core/characterisation-tests/helpers/processMessageCore.ts
@@ -9,7 +9,7 @@ import type { ProcessMessageOptions } from "./processMessage"
 
 import MockPncGateway from "../../comparison/lib/MockPncGateway"
 import parseIncomingMessage from "../../comparison/lib/parseIncomingMessage"
-import CoreAuditLogger from "../../lib/CoreAuditLogger"
+import CoreAuditLogger from "../../lib/auditLog/CoreAuditLogger"
 import { PncApiError } from "../../lib/pnc/PncGateway"
 import CorePhase1 from "../../phase1/phase1"
 import CorePhase2 from "../../phase2/phase2"
