+++ packages/core/comparison/lib/comparePhase1.ts
@@ -8,7 +8,7 @@ import type { OldPhase1Comparison, Phase1Comparison } from "../types/ComparisonF
 import type ComparisonResultDetail from "../types/ComparisonResultDetail"
 import type { ComparisonResultDebugOutput } from "../types/ComparisonResultDetail"
 
-import CoreAuditLogger from "../../lib/CoreAuditLogger"
+import CoreAuditLogger from "../../lib/auditLog/CoreAuditLogger"
 import { extractExceptionsFromXml, parseAhoXml } from "../../lib/parse/parseAhoXml"
 import serialiseToXml from "../../lib/serialise/ahoXml/serialiseToXml"
 import getMessageType from "../../phase1/lib/getMessageType"
