+++ packages/core/phase1/tests/helpers/processMessage.ts
@@ -7,7 +7,7 @@ import type { ResultedCaseMessageParsedXml } from "../../../types/SpiResult"
 import type Phase1Result from "../../types/Phase1Result"
 
 import MockPncGateway from "../../../comparison/lib/MockPncGateway"
-import CoreAuditLogger from "../../../lib/CoreAuditLogger"
+import CoreAuditLogger from "../../../lib/auditLog/CoreAuditLogger"
 import extractExceptionsFromAho from "../../../lib/parse/parseAhoXml/extractExceptionsFromXml"
 import parseSpiResult from "../../../lib/parse/parseSpiResult"
 import transformSpiToAho from "../../../lib/parse/transformSpiToAho"
