+++ packages/core/scripts/testServer.ts
@@ -7,7 +7,7 @@ import type { PncQueryResult } from "../types/PncQueryResult"
 
 import MockPncGateway from "../comparison/lib/MockPncGateway"
 import parseIncomingMessage from "../comparison/lib/parseIncomingMessage"
-import CoreAuditLogger from "../lib/CoreAuditLogger"
+import CoreAuditLogger from "../lib/auditLog/CoreAuditLogger"
 import CorePhase1 from "../phase1/phase1"
 import CorePhase2 from "../phase2/phase2"
 import Phase from "../types/Phase"
