+++ packages/ui/test/services/storeAuditLogEvents.integration.test.ts
@@ -2,7 +2,7 @@ import type { AuditLogEvent } from "@moj-bichard7/common/types/AuditLogEvent"
 import { auditLogEventLookup } from "@moj-bichard7/common/types/AuditLogEvent"
 import EventCategory from "@moj-bichard7/common/types/EventCategory"
 import EventCode from "@moj-bichard7/common/types/EventCode"
-import getAuditLogEvent from "@moj-bichard7/core/lib/getAuditLogEvent"
+import getAuditLogEvent from "@moj-bichard7/core/lib/auditLog/getAuditLogEvent"
 import axios from "axios"
 import { randomUUID } from "crypto"
 import getDataSource from "services/getDataSource"
